+++ labs/lab03/backend/main.go
@@ -1,21 +1,28 @@
 package main
 
 import (
+	"lab03-backend/api"
+	"lab03-backend/storage"
 	"log"
+	"net/http"
+	"time"
 )
 
 func main() {
-	// TODO: Create a new memory storage instance
-	// TODO: Create a new API handler with the storage
-	// TODO: Setup routes using the handler
-	// TODO: Configure server with:
-	//   - Address: ":8080"
-	//   - Handler: the router
-	//   - ReadTimeout: 15 seconds
-	//   - WriteTimeout: 15 seconds
-	//   - IdleTimeout: 60 seconds
-	// TODO: Add logging to show server is starting
-	// TODO: Start the server and handle any errors
+	memoryStorage := storage.NewMemoryStorage()
+	handler := api.NewHandler(memoryStorage)
+	router := handler.SetupRoutes()
 
-	log.Println("TODO: Implement main function")
+	server := &http.Server{
+		Addr:         ":8080",
+		Handler:      router,
+		ReadTimeout:  15 * time.Second,
+		WriteTimeout: 15 * time.Second,
+		IdleTimeout:  60 * time.Second,
+	}
+
+	log.Println("Server starting on :8080")
+	if err := server.ListenAndServe(); err != nil {
+		log.Fatal("Server failed to start:", err)
+	}
 }
