+++ labs/lab03/backend/storage/memory.go
@@ -3,79 +3,90 @@ package storage
 import (
 	"errors"
 	"lab03-backend/models"
+	"sync"
 )
 
 // MemoryStorage implements in-memory storage for messages
 type MemoryStorage struct {
-	// TODO: Add mutex field for thread safety (sync.RWMutex)
-	// TODO: Add messages field as map[int]*models.Message
-	// TODO: Add nextID field of type int for auto-incrementing IDs
+	mu       sync.RWMutex
+	messages map[int]*models.Message
+	nextID   int
 }
 
 // NewMemoryStorage creates a new in-memory storage instance
 func NewMemoryStorage() *MemoryStorage {
-	// TODO: Return a new MemoryStorage instance with initialized fields
-	// Initialize messages as empty map
-	// Set nextID to 1
-	return nil
+	return &MemoryStorage{
+		messages: make(map[int]*models.Message),
+		nextID:   1,
+	}
 }
 
 // GetAll returns all messages
 func (ms *MemoryStorage) GetAll() []*models.Message {
-	// TODO: Implement GetAll method
-	// Use read lock for thread safety
-	// Convert map values to slice
-	// Return slice of all messages
-	return nil
+	ms.mu.RLock()
+	defer ms.mu.RUnlock()
+
+	messages := make([]*models.Message, 0, len(ms.messages))
+	for _, msg := range ms.messages {
+		messages = append(messages, msg)
+	}
+	return messages
 }
 
 // GetByID returns a message by its ID
 func (ms *MemoryStorage) GetByID(id int) (*models.Message, error) {
-	// TODO: Implement GetByID method
-	// Use read lock for thread safety
-	// Check if message exists in map
-	// Return message or error if not found
-	return nil, nil
+	ms.mu.RLock()
+	defer ms.mu.RUnlock()
+
+	msg, exists := ms.messages[id]
+	if !exists {
+		return nil, ErrMessageNotFound
+	}
+	return msg, nil
 }
 
 // Create adds a new message to storage
 func (ms *MemoryStorage) Create(username, content string) (*models.Message, error) {
-	// TODO: Implement Create method
-	// Use write lock for thread safety
-	// Get next available ID
-	// Create new message using models.NewMessage
-	// Add message to map
-	// Increment nextID
-	// Return created message
-	return nil, nil
+	ms.mu.Lock()
+	defer ms.mu.Unlock()
+
+	id := ms.nextID
+	msg := models.NewMessage(id, username, content)
+	ms.messages[id] = msg
+	ms.nextID++
+	return msg, nil
 }
 
 // Update modifies an existing message
 func (ms *MemoryStorage) Update(id int, content string) (*models.Message, error) {
-	// TODO: Implement Update method
-	// Use write lock for thread safety
-	// Check if message exists
-	// Update the content field
-	// Return updated message or error if not found
-	return nil, nil
+	ms.mu.Lock()
+	defer ms.mu.Unlock()
+
+	msg, exists := ms.messages[id]
+	if !exists {
+		return nil, ErrMessageNotFound
+	}
+	msg.Content = content
+	return msg, nil
 }
 
 // Delete removes a message from storage
 func (ms *MemoryStorage) Delete(id int) error {
-	// TODO: Implement Delete method
-	// Use write lock for thread safety
-	// Check if message exists
-	// Delete from map
-	// Return error if message not found
+	ms.mu.Lock()
+	defer ms.mu.Unlock()
+
+	if _, exists := ms.messages[id]; !exists {
+		return ErrMessageNotFound
+	}
+	delete(ms.messages, id)
 	return nil
 }
 
 // Count returns the total number of messages
 func (ms *MemoryStorage) Count() int {
-	// TODO: Implement Count method
-	// Use read lock for thread safety
-	// Return length of messages map
-	return 0
+	ms.mu.RLock()
+	defer ms.mu.RUnlock()
+	return len(ms.messages)
 }
 
 // Common errors
