+++ labs/lab03/frontend/lib/services/api_service.dart
@@ -1,100 +1,221 @@
+import 'dart:async';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
+import 'package:http/http.dart';
 import '../models/message.dart';
 
 class ApiService {
-  // TODO: Add static const String baseUrl = 'http://localhost:8080';
-  // TODO: Add static const Duration timeout = Duration(seconds: 30);
-  // TODO: Add late http.Client _client field
+  static const String baseUrl = 'http://localhost:8080';
+  static const Duration timeout = Duration(seconds: 30);
+  late http.Client _client;
 
-  // TODO: Add constructor that initializes _client = http.Client();
+  ApiService({http.Client? client}) {
+    _client = client ?? http.Client();
+  }
 
-  // TODO: Add dispose() method that calls _client.close();
+  void dispose() {
+    _client.close();
+  }
 
-  // TODO: Add _getHeaders() method that returns Map<String, String>
-  // Return headers with 'Content-Type': 'application/json' and 'Accept': 'application/json'
+  Map<String, String> _getHeaders() {
+    return {
+      'Content-Type': 'application/json',
+      'Accept': 'application/json',
+    };
+  }
 
-  // TODO: Add _handleResponse<T>() method with parameters:
-  // http.Response response, T Function(Map<String, dynamic>) fromJson
-  // Check if response.statusCode is between 200-299
-  // If successful, decode JSON and return fromJson(decodedData)
-  // If 400-499, throw client error with message from response
-  // If 500-599, throw server error
-  // For other status codes, throw general error
+  T _handleResponse<T>(
+      http.Response response, T Function(Map<String, dynamic>) fromJson) {
+    if (response.statusCode >= 200 && response.statusCode < 300) {
+      final data = jsonDecode(response.body);
+      return fromJson(data);
+    } else if (response.statusCode >= 400 && response.statusCode < 500) {
+      throw ApiException('Client error: ${response.body}');
+    } else if (response.statusCode >= 500) {
+      throw ServerException('Server error: ${response.body}');
+    } else {
+      throw ApiException('Unexpected error: ${response.statusCode}');
+    }
+  }
 
-  // Get all messages
   Future<List<Message>> getMessages() async {
-    // TODO: Implement getMessages
-    // Make GET request to '$baseUrl/api/messages'
-    // Use _handleResponse to parse response into List<Message>
-    // Handle network errors and timeouts
-    throw UnimplementedError('TODO: Implement getMessages');
+    try {
+      final response = await _client
+          .get(
+            Uri.parse('$baseUrl/api/messages'),
+            headers: _getHeaders(),
+          )
+          .timeout(timeout);
+
+      return _handleResponse(response, (data) {
+        if (data is List) {
+          return (data as List<dynamic>)
+              .map((item) => Message.fromJson(item as Map<String, dynamic>))
+              .toList();
+        } else if (data.containsKey('success') && data.containsKey('data')) {
+          final apiResponse =
+              ApiResponse<List<dynamic>>.fromJson(data, (d) => d);
+          return (apiResponse.data as List<dynamic>)
+              .map((item) => Message.fromJson(item as Map<String, dynamic>))
+              .toList();
+        } else {
+          throw ApiException('Unexpected response format');
+        }
+      });
+    } catch (e) {
+      throw NetworkException('Network error: $e');
+    }
   }
 
-  // Create a new message
   Future<Message> createMessage(CreateMessageRequest request) async {
-    // TODO: Implement createMessage
-    // Validate request using request.validate()
-    // Make POST request to '$baseUrl/api/messages'
-    // Include request.toJson() in body
-    // Use _handleResponse to parse response
-    // Extract message from ApiResponse.data
-    throw UnimplementedError('TODO: Implement createMessage');
-  }
+  final error = request.validate();
+  if (error != null) throw ValidationException(error);
 
-  // Update an existing message
-  Future<Message> updateMessage(int id, UpdateMessageRequest request) async {
-    // TODO: Implement updateMessage
-    // Validate request using request.validate()
-    // Make PUT request to '$baseUrl/api/messages/$id'
-    // Include request.toJson() in body
-    // Use _handleResponse to parse response
-    // Extract message from ApiResponse.data
-    throw UnimplementedError('TODO: Implement updateMessage');
-  }
+  try {
+    final response = await _client.post(
+      Uri.parse('$baseUrl/api/messages'),
+      headers: _getHeaders(),
+      body: jsonEncode(request.toJson()),
+    ).timeout(timeout);
 
-  // Delete a message
-  Future<void> deleteMessage(int id) async {
-    // TODO: Implement deleteMessage
-    // Make DELETE request to '$baseUrl/api/messages/$id'
-    // Check if response.statusCode is 204
-    // Throw error if deletion failed
-    throw UnimplementedError('TODO: Implement deleteMessage');
-  }
-
-  // Get HTTP status information
-  Future<HTTPStatusResponse> getHTTPStatus(int statusCode) async {
-    // TODO: Implement getHTTPStatus
-    // Make GET request to '$baseUrl/api/status/$statusCode'
-    // Use _handleResponse to parse response
-    // Extract HTTPStatusResponse from ApiResponse.data
-    throw UnimplementedError('TODO: Implement getHTTPStatus');
-  }
-
-  // Health check
-  Future<Map<String, dynamic>> healthCheck() async {
-    // TODO: Implement healthCheck
-    // Make GET request to '$baseUrl/api/health'
-    // Return decoded JSON response
-    throw UnimplementedError('TODO: Implement healthCheck');
+    return _handleResponse(response, (data) {
+      if (data.containsKey('id')) return Message.fromJson(data);
+      if (data.containsKey('success') && data.containsKey('data')) {
+        return Message.fromJson(data['data']);
+      }
+      throw ApiException('Unexpected response format');
+    });
+  } on http.ClientException catch (e) {
+    throw NetworkException('Network error: ${e.message}');
+  } on ApiException {
+    rethrow;
+  } catch (e) {
+    throw NetworkException('Network error: $e');
   }
 }
 
-// Custom exceptions
+  Future<Message> updateMessage(int id, UpdateMessageRequest request) async {
+    final validationError = request.validate();
+    if (validationError != null) {
+      throw ValidationException(validationError);
+    }
+
+    try {
+      final uri = Uri.parse('$baseUrl/api/messages/$id');
+      final response = await _client
+          .put(
+            uri,
+            headers: _getHeaders(),
+            body: jsonEncode(request.toJson()),
+          )
+          .timeout(timeout);
+
+      return _handleResponse(response, (data) {
+      if (data.containsKey('id')) return Message.fromJson(data);
+      if (data.containsKey('success') && data.containsKey('data')) {
+        return Message.fromJson(data['data'] as Map<String, dynamic>);
+      }
+      throw ApiException('Unexpected response format');
+    });
+  } on http.ClientException catch (e) {
+    throw NetworkException('Network error: ${e.message}');
+  } on ApiException {
+    rethrow;
+  } catch (e) {
+    throw NetworkException('Network error: $e');
+  }
+}
+
+  Future<void> deleteMessage(int id) async {
+    try {
+      final uri = Uri.parse('$baseUrl/api/messages/$id');
+      final response = await _client
+          .delete(
+            uri,
+            headers: _getHeaders(),
+          )
+          .timeout(timeout);
+
+      if (response.statusCode != 204) {
+        throw ApiException('Failed to delete message: ${response.statusCode}');
+      }
+    } catch (e) {
+      if (e is http.ClientException) {
+        throw NetworkException('Network error: ${e.message}');
+      }
+      rethrow;
+    }
+  }
+
+  Future<HTTPStatusResponse> getHTTPStatus(int statusCode) async {
+    if (statusCode < 100 || statusCode >= 600) {
+      throw ValidationException('Invalid status code: $statusCode');
+    }
+
+    try {
+      final uri = Uri.parse('$baseUrl/api/status/$statusCode');
+      final response = await _client
+          .get(
+            uri,
+            headers: _getHeaders(),
+          )
+          .timeout(timeout);
+
+      return _handleResponse(response, (data) {
+  final responseData = (data.containsKey('success') && data.containsKey('data'))
+      ? data['data'] as Map<String, dynamic>
+      : data;
+  
+  if (!responseData.containsKey('status_code')) {
+    throw ApiException('Invalid response format: missing status_code');
+  }
+  
+  return HTTPStatusResponse.fromJson(responseData);
+});
+    } catch (e) {
+      if (e is ApiException) rethrow;
+      throw NetworkException('Network error: $e');
+    }
+  }
+
+  Future<Map<String, dynamic>> healthCheck() async {
+    try {
+      final uri = Uri.parse('$baseUrl/api/health');
+      final response = await _client
+          .get(
+            uri,
+            headers: _getHeaders(),
+          )
+          .timeout(timeout);
+
+      if (response.statusCode == 200) {
+        return jsonDecode(response.body);
+      } else {
+        throw ApiException('Health check failed: ${response.body}');
+      }
+    } catch (e) {
+      if (e is ApiException) rethrow;
+      throw NetworkException('Network error: $e');
+    }
+  }
+}
+
 class ApiException implements Exception {
-  // TODO: Add final String message field
-  // TODO: Add constructor ApiException(this.message);
-  // TODO: Override toString() to return 'ApiException: $message'
+  final String message;
+  ApiException(this.message);
+
+  @override
+  String toString() => 'ApiException: $message';
 }
 
 class NetworkException extends ApiException {
-  // TODO: Add constructor NetworkException(String message) : super(message);
+  NetworkException(String message) : super(message);
 }
 
 class ServerException extends ApiException {
-  // TODO: Add constructor ServerException(String message) : super(message);
+  ServerException(String message) : super(message);
 }
 
 class ValidationException extends ApiException {
-  // TODO: Add constructor ValidationException(String message) : super(message);
+  ValidationException(String message) : super(message);
 }
