--- labs/lab03/lab03/backend/go.mod
@@ -1,5 +1,0 @@
-module lab03-backend
-
-go 1.24
-
-require github.com/gorilla/mux v1.8.0
