--- labs/lab03/lab03/backend/models/message.go
@@ -1,68 +1,0 @@
-package models
-
-import (
-	"errors"
-	"time"
-)
-
-// Message represents a chat message
-type Message struct {
-	ID        int       `json:"id"`
-	Username  string    `json:"username"`
-	Content   string    `json:"content"`
-	Timestamp time.Time `json:"timestamp"`
-}
-
-// CreateMessageRequest represents the request to create a new message
-type CreateMessageRequest struct {
-	Username string `json:"username" validate:"required"`
-	Content  string `json:"content" validate:"required"`
-}
-
-// UpdateMessageRequest represents the request to update a message
-type UpdateMessageRequest struct {
-	Content string `json:"content" validate:"required"`
-}
-
-// HTTPStatusResponse represents the response for HTTP status code endpoint
-type HTTPStatusResponse struct {
-	StatusCode  int    `json:"status_code"`
-	ImageURL    string `json:"image_url"`
-	Description string `json:"description"`
-}
-
-// APIResponse represents a generic API response
-type APIResponse struct {
-	Success bool        `json:"success"`
-	Data    interface{} `json:"data,omitempty"`
-	Error   string      `json:"error,omitempty"`
-}
-
-// NewMessage creates a new message with the current timestamp
-func NewMessage(id int, username, content string) *Message {
-	return &Message{
-		ID:        id,
-		Username:  username,
-		Content:   content,
-		Timestamp: time.Now(),
-	}
-}
-
-// Validate checks if the create message request is valid
-func (r *CreateMessageRequest) Validate() error {
-	if r.Username == "" {
-		return errors.New("username is required")
-	}
-	if r.Content == "" {
-		return errors.New("content is required")
-	}
-	return nil
-}
-
-// Validate checks if the update message request is valid
-func (r *UpdateMessageRequest) Validate() error {
-	if r.Content == "" {
-		return errors.New("content is required")
-	}
-	return nil
-}
