--- labs/lab03/lab03/frontend/lib/screens/chat_screen.dart
@@ -1,238 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import '../models/message.dart';
-import '../services/api_service.dart';
-import '../main.dart';
-
-class ChatScreen extends StatefulWidget {
-  const ChatScreen({super.key});
-
-  @override
-  State<ChatScreen> createState() => _ChatScreenState();
-}
-
-class _ChatScreenState extends State<ChatScreen> {
-  final TextEditingController _usernameController = TextEditingController();
-  final TextEditingController _messageController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      context.read<ChatProvider>().loadMessages();
-    });
-  }
-
-  @override
-  void dispose() {
-    _usernameController.dispose();
-    _messageController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _sendMessage() async {
-    if (_usernameController.text.isEmpty || _messageController.text.isEmpty) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Please fill in all fields')),
-      );
-      return;
-    }
-
-    final request = CreateMessageRequest(
-      username: _usernameController.text,
-      content: _messageController.text,
-    );
-
-    await context.read<ChatProvider>().createMessage(request);
-    _messageController.clear();
-
-    if (mounted && context.read<ChatProvider>().error == null) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Message sent!')),
-      );
-    }
-  }
-
-  Future<void> _showHTTPStatus(int statusCode) async {
-    try {
-      final apiService = context.read<ApiService>();
-      final status = await apiService.getHTTPStatus(statusCode);
-
-      if (mounted) {
-        showDialog(
-          context: context,
-          builder: (context) => AlertDialog(
-            title: Text('HTTP Status: ${status.statusCode}'),
-            content: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                Text(status.description),
-                const SizedBox(height: 16),
-                Image.network(
-                  status.imageUrl,
-                  height: 200,
-                  errorBuilder: (context, error, stackTrace) =>
-                      const Text('Failed to load image'),
-                ),
-              ],
-            ),
-            actions: [
-              TextButton(
-                onPressed: () => Navigator.of(context).pop(),
-                child: const Text('Close'),
-              ),
-            ],
-          ),
-        );
-      }
-    } catch (e) {
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text('Error: $e')),
-        );
-      }
-    }
-  }
-
-  Widget _buildMessageTile(Message message) {
-    return ListTile(
-      leading: CircleAvatar(
-        child: Text(message.username.isNotEmpty
-            ? message.username[0].toUpperCase()
-            : '?'),
-      ),
-      title: Text(
-          '${message.username} • ${message.timestamp.toString().substring(0, 16)}'),
-      subtitle: Text(message.content),
-      onTap: () => _showHTTPStatus([200, 404, 500][message.id % 3]),
-    );
-  }
-
-  Widget _buildMessageInput() {
-    return Container(
-      padding: const EdgeInsets.all(16),
-      color: Colors.grey[100],
-      child: Column(
-        children: [
-          TextField(
-            controller: _usernameController,
-            decoration: const InputDecoration(
-              hintText: 'Enter your username',
-              border: OutlineInputBorder(),
-            ),
-          ),
-          const SizedBox(height: 8),
-          TextField(
-            controller: _messageController,
-            decoration: const InputDecoration(
-              hintText: 'Enter your message',
-              border: OutlineInputBorder(),
-            ),
-          ),
-          const SizedBox(height: 8),
-          Row(
-            children: [
-              ElevatedButton(
-                onPressed: _sendMessage,
-                child: const Text('Send'),
-              ),
-              const SizedBox(width: 8),
-              ElevatedButton(
-                onPressed: () => _showHTTPStatus(200),
-                child: const Text('200 OK'),
-              ),
-              const SizedBox(width: 8),
-              ElevatedButton(
-                onPressed: () => _showHTTPStatus(404),
-                child: const Text('404 Not Found'),
-              ),
-              const SizedBox(width: 8),
-              ElevatedButton(
-                onPressed: () => _showHTTPStatus(500),
-                child: const Text('500 Error'),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildErrorWidget(String error) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          const Icon(Icons.error_outline, color: Colors.red, size: 64),
-          const SizedBox(height: 16),
-          Text(error, style: const TextStyle(color: Colors.red)),
-          const SizedBox(height: 16),
-          ElevatedButton(
-            onPressed: () => context.read<ChatProvider>().loadMessages(),
-            child: const Text('Retry'),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildLoadingWidget() {
-    return const Center(child: CircularProgressIndicator());
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('REST API Chat'),
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.refresh),
-            onPressed: () => context.read<ChatProvider>().refreshMessages(),
-          ),
-        ],
-      ),
-      body: Column(
-        children: [
-          Expanded(
-            child: Consumer<ChatProvider>(
-              builder: (context, provider, child) {
-                if (provider.isLoading) {
-                  return _buildLoadingWidget();
-                } else if (provider.error != null) {
-                  return _buildErrorWidget(provider.error!);
-                } else if (provider.messages.isEmpty) {
-                  return const Center(
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Icon(Icons.chat_bubble_outline,
-                            size: 64, color: Colors.grey),
-                        SizedBox(height: 16),
-                        Text('No messages yet'),
-                        Text('Send your first message to get started!'),
-                      ],
-                    ),
-                  );
-                } else {
-                  return ListView.builder(
-                    itemCount: provider.messages.length,
-                    itemBuilder: (context, index) {
-                      return _buildMessageTile(provider.messages[index]);
-                    },
-                  );
-                }
-              },
-            ),
-          ),
-          _buildMessageInput(),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () => context.read<ChatProvider>().refreshMessages(),
-        child: const Icon(Icons.refresh),
-      ),
-    );
-  }
-}
-
