--- labs/lab03/lab03/frontend/pubspec.yaml
@@ -1,23 +1,0 @@
-name: lab03_frontend
-description: Lab 03 REST API Chat Frontend
-
-publish_to: 'none'
-
-version: 1.0.0+1
-
-environment:
-  sdk: '>=3.0.0 <4.0.0'
-
-dependencies:
-  flutter:
-    sdk: flutter
-  http: ^1.1.0
-  provider: ^6.1.1
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^3.0.1
-
-flutter:
-  uses-material-design: true
